requirements.txt
sqlalchemy
flask
docopt
bcrypt
flask-restplus
wget
ipy
python-dateutil
netifaces
flask-cors
